% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ulanrdb.R
\name{build_ulanrdb}
\alias{build_ulanrdb}
\title{Build a local ULAN database}
\usage{
build_ulanrdb()
}
\description{
For use with method = "local" options on the ulan_ functions. This downloads
a current database of IDs, preferred/alternate names, and biographical
details from \url{http://vocab.getty.edu/sparql}. This requires that the
container package \code{ulanrdb} is installed.
}

