% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ulanrdata.R
\name{check_ulanrdata_package}
\alias{check_ulanrdata_package}
\title{Check whether to install data for local ULAN matching, and install if
necessary}
\usage{
check_ulanrdata_package()
}
\description{
If the ulanrdata package is not installed, install it from GitHub using
devtools. If it is not up to date, reinstall it.
}

