% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_results.R
\name{construct_results}
\alias{construct_results}
\title{Construct a dataframe}
\usage{
construct_results(results, name = NULL)
}
\arguments{
\item{results}{A dataframe, or NA. If NA, this function will return a warning
and an empty data frame with the same set of columns as a complete result.}

\item{name}{The name for which these results are being returned. This is only
used in the case of NA results to throw a warning.}
}
\value{
A data frame
}
\description{
Construct a dataframe
}

