% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{date_filter}
\alias{date_filter}
\title{Construct SPARQL date filter}
\usage{
date_filter(inclusive, early_year, late_year)
}
\description{
Constructs a portion of the SPARQL query to filter artists based on life
dates
}

