% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ulanrdata.R
\name{install_ulanrdata_package}
\alias{install_ulanrdata_package}
\title{Install the ulanrdata package after checking with the user}
\usage{
install_ulanrdata_package()
}
\description{
Install the ulanrdata package after checking with the user
}

