% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{sparql_url}
\alias{sparql_url}
\title{Format a SPARQL query as a URL}
\usage{
sparql_url(query)
}
\arguments{
\item{query}{Passed by ulan_sparql_match_handler.}
}
\description{
Properly escapes the query to send to the Getty SPARQL endpoint
}

