% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulanr.R
\name{ulan_data}
\alias{ulan_data}
\title{Return a dataframe of matching ULAN IDs with attributes}
\usage{
ulan_data(names, early_year = -9999, late_year = 2090, inclusive = TRUE,
  method = c("sparql", "local"), max_results = 1)
}
\arguments{
\item{names}{A character vector of names to match to a canonical ULAN id.}

\item{early_year}{Match only artists who died after this year. Like
\code{late_year}, this argument should be a numeric vector of length 1, or
of the same length as \code{names}. If length 1, the same date restrictions
will be used to match every value of \code{names}. Otherwise, each name
match can be restricted to its own pair of early_year and late_year. If no
\code{early_year} or \code{late_year} are specified, then artists from all
time periods will be eligible for matching. Any NA values in
\code{early_year} or \code{late_year} will be coerced to default maxima and
minima.}

\item{late_year}{Match only artists who were born before this year.}

\item{inclusive}{Method for filtering search results using the
early_year/late_year parameters. TRUE (the default) will only include
artists whose life dates fall within the range [late_year, early_year].
FALSE will include artists whose life dates intersect with [early_year,
late_year]}

\item{method}{This value determines which method will be used to match the
name to a canonical ULAN id. \code{sparql} will query the Getty's live
endpoint, relying on its Lucene index for finding close matches, while
\code{local} instead uses string distance measures based on a local table
of ULAN entries.}

\item{max_results}{The maximum number of results to return. Defaults to 5.
Depending on the query, the actual number of results returned may be lower.
When \code{method = "sparql"} is used, values over 50 will be ignored.}
}
\value{
A data frame with 6 columns:
\describe{
\item{\code{id}}{integer. ULAN id}
\item{\code{pref_name}}{character. ULAN preferred name}
\item{\code{birth_year}}{integer. Artist birth year, if assigned.}
\item{\code{death_year}}{integer. Artist death year, if assigned}
\item{\code{gender}}{character. Artist gender, if assigned.}
\item{\code{nationality}}{character. Artist nationality, if assigned.}
\item{\code{score}}{numeric. The score of the result. When \code{method =
"sparql"}, this is the Lucene index score (a higher score for a closer
match). When \code{method = "local"}, it will instead be a scaled string
distance score.}
\item{\code{names}}{character. The input vector.}
}
}
\description{
This function is a wrapper around \link{ulan_match} that simply returns a
data.frame with the top match ID for each name given (as if \code{max_results} were
1)
}
\note{
\code{max_results} will be ignored.
}

