% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulanr.R
\name{ulan_id}
\alias{ulan_id}
\title{Name to ULAN ID}
\usage{
ulan_id(name, years = NULL, method = c("sparql"))
}
\arguments{
\item{name}{A name to match with a canonical ULAN id.}

\item{years}{Years used to restrict the matching functions. This argument
should be a range of years in the form \code{c(1550, 1650)}. ULAN IDs will
only be looked up for artists whose life dates intersect with this range.
If no years are specified, then ULAN artists will be matched from all
periods.)}

\item{method}{This value determines which method will be used to match the
  name to a canonical ULAN id.}
}
\value{
A canonical ULAN id for an artist as a 9-digit integer. If no match
  is found, then NULL will be returned isntead.
}
\description{
Queries the Getty ULAN to find the best matching ID for a given string. You
may filter the results by specifying an early or late date.
}
\note{
\code{method = "sparql"} requires an internet connection.
}
\examples{
\dontrun{ulan_id("Rembrandt", years = c(1600, 1700), method = "sparql")}
\dontrun{ulan_id("Rembrandt", years = c(1700, 1800), method = "sparql")}
}

