% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{ulan_sparql}
\alias{ulan_sparql}
\title{Search for a matching ULAN id by using the Getty's live SPARQL endpoint and
its Lucene index.}
\usage{
ulan_sparql(name, years)
}
\arguments{
\item{name}{A character string of an artist's name}

\item{years}{Window in which to search for matching artists}
}
\description{
This internal function implements the \code{method = "sparql"} option for
\link{ulan_id}. See that funciton for documentation.
}

