% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{ulan_sparql_data_handler}
\alias{ulan_sparql_data_handler}
\title{Search for a matching ULAN id and its associated databy using the Getty's
live SPARQL endpoint and its Lucene index.}
\usage{
ulan_sparql_data_handler(name, early_year, late_year, inclusive)
}
\arguments{
\item{name}{A character string of an artist's name}

\item{early_year}{Match only artists who died after this year.}

\item{late_year}{Match only artists who were born before this year.}

\item{inclusive}{Logical. Should life dates be filtered inclusive of the
[early_year, late_year] range?}
}
\description{
This internal function implements the \code{method = "sparql"} option for
\link{ulan_data}. See that funciton for documentation.
}

