% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{ulan_sparql_handler}
\alias{ulan_sparql_handler}
\title{Search for a matching ULAN id by using the Getty's live SPARQL endpoint and
its Lucene index.}
\usage{
ulan_sparql_handler(name, early_year, late_year)
}
\arguments{
\item{name}{A character string of an artist's name}

\item{early_year}{Match only artists who died after this year.}

\item{late_year}{Match only artists who were born before this year.}
}
\description{
This internal function implements the \code{method = "sparql"} option for
\link{ulan_id}. See that funciton for documentation.
}

