% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{ulan_sparql_id}
\alias{ulan_sparql_id}
\title{Iterate the SPARQL method over the provided vectors}
\usage{
ulan_sparql_id(names, early_year, late_year, inclusive, progress_bar)
}
\arguments{
\item{names}{A character string of an artist's name}

\item{early_year}{Match only artists who died after this year.}

\item{late_year}{Match only artists who were born before this year.}

\item{inclusive}{Logical. Should life dates be filtered inclusive of the
[early_year, late_year] range?}

\item{progress_bar}{Display a progress bar for long vectors.}
}
\description{
This internal function maps the inputs from the generic \link{ulan_id}
function to the SPARQL implementation
}

