% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{ulan_sparql_match}
\alias{ulan_sparql_match}
\title{Iterate the SPARQL method over the provided vectors}
\usage{
ulan_sparql_match(names, early_year, late_year, inclusive, max_results)
}
\description{
This internal function maps the inputs from the generic \link{ulan_data}
function to the SPARQL implementation
}

