% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{ulan_sparql_match_handler}
\alias{ulan_sparql_match_handler}
\title{Search for a matching ULAN id and its associated databy using the Getty's
live SPARQL endpoint and its Lucene index.}
\usage{
ulan_sparql_match_handler(name, early_year, late_year, inclusive, max_results,
  ...)
}
\arguments{
\item{name}{Passed from ulan_match.}

\item{early_year}{Passed from ulan_match.}

\item{late_year}{Passed from ulan_match.}

\item{inclusive}{Passed from ulan_match.}

\item{max_results}{Passed from ulan_match.}

\item{...}{ignores stringdist_ops argument, if present.}
}
\description{
This internal function implements the \code{method = "sparql"} option for
\link{ulan_data}. See that funciton for documentation.
}

