% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulan_stringdist.R
\name{ulan_stringdist_lookup}
\alias{ulan_stringdist_lookup}
\title{String Distance Lookup}
\usage{
ulan_stringdist_lookup(name, early_year, late_year, inclusive, stringdist_ops)
}
\arguments{
\item{name}{Passed from ulan_stringdist_match_handler.}

\item{early_year}{Passed from ulan_stringdist_match_handler.}

\item{late_year}{Passed from ulan_stringdist_match_handler.}

\item{inclusive}{Passed from ulan_stringdist_match_handler.}

\item{stringdist_ops}{Passed from ulan_stringdist_match_handler, to be used
in stringdist() call.}
}
\description{
Run the table lookup using stringdist function, returning a data frame of all matches
}

