% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulan_stringdist.R
\name{ulan_stringdist_match_handler}
\alias{ulan_stringdist_match_handler}
\title{Handle strindist results}
\usage{
ulan_stringdist_match_handler(name, early_year, late_year, inclusive,
  max_results, stringdist_ops = NULL)
}
\arguments{
\item{name}{Passed from ulan_match.}

\item{early_year}{Passed from ulan_match.}

\item{late_year}{Passed from ulan_match.}

\item{inclusive}{Passed from ulan_match.}

\item{max_results}{Passed from ulan_match.}

\item{stringdist_ops}{Passed from ulan_match.}
}
\description{
Returns one-row data frame with artist attributes
}

