% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulanr-package.R
\docType{package}
\name{ulanr}
\alias{ulanr}
\alias{ulanr-package}
\title{ulanr: Reconcile Names to the Getty Union List of Artist Names}
\description{
Matches variant spellings of artist names to canonical ids by
calling the Getty Union List of Artist Names SPARQL endpoint.
}

