% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulanr.R
\name{validate_input}
\alias{validate_input}
\title{Validate input variables}
\usage{
validate_input(names, early_year, late_year)
}
\arguments{
\item{names}{A character vector of names to match to a canonical ULAN id.}

\item{early_year}{Match only artists who died after this year.}

\item{late_year}{Match only artists who were born before this year.}
}
\description{
A helper function that validates the classes and lengths of inputs to
uland_id and ulan_data functions.  Will stop() code if there are any invalid
variable types
}

