% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{date_filter}
\alias{date_filter}
\title{Construct SPARQL date filter}
\usage{
date_filter(inclusive, early_year, late_year)
}
\arguments{
\item{inclusive}{Passed by ulan_sparql_match_handler.}

\item{early_year}{Passed by ulan_sparql_match_handler.}

\item{late_year}{Passed by ulan_sparql_match_handler.}
}
\description{
Constructs a portion of the SPARQL query to filter artists based on life
dates
}

