% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{date_filter}
\alias{date_filter}
\title{Construct SPARQL date filter}
\usage{
date_filter(inclusive, early_year, late_year)
}
\arguments{
\item{inclusive}{Logical. Should life dates be filtered inclusive of the
[early_year, late_year] range?}

\item{early_year}{Match only artists who died after this year.}

\item{late_year}{Match only artists who were born before this year.}
}
\description{
Constructs a portion of the SPARQL query to filter artists based on life
dates
}

