% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulanr.R
\name{ulan_data}
\alias{ulan_data}
\title{Name to ULAN data}
\usage{
ulan_data(names, early_year = -9999, late_year = 2090, inclusive = TRUE,
  method = c("sparql"), progress_bar = "default")
}
\arguments{
\item{names}{A character vector of names to match to a canonical ULAN id.}

\item{early_year}{Match only artists who died after this year. Like
\code{late_year}, this argument should be a numeric vector of length 1, or
of the same length as \code{names}. If length 1, the same date restrictions
will be used to match every value of \code{names}. Otherwise, each name
match can be restricted to its own pair of early_year and late_year. If no
\code{early_year} or \code{late_year} are specified, then artists from all
time periods will be eligible for matching. Any NA values in
\code{early_year} or \code{late_year} will be coerced to default maxima and
minima.}

\item{late_year}{Match only artists who were born before this year.}

\item{inclusive}{Method for filtering search results using the
early_year/late_year parameters. TRUE (the default) will only include
artists whose life dates fall within the range [late_year, early_year].
FALSE will include artists whose life dates intersect with [early_year,
late_year]}

\item{method}{This value determines which method will be used to match the
name to a canonical ULAN id.}

\item{progress_bar}{Display or hide a progress bar. By default, will only
  display a progress bar if processing more than 50 names. Passing TRUE will
  force a progress bar to appear; passing FALSE will always hide it.}
}
\value{
A data frame with 6 columns.
\describe{
\item{\code{name}}{Original input vector}
\item{\code{id}}{ULAN id}
\item{\code{pref_name}}{ULAN preferred name}
\item{\code{startdate}}{Artist birth date}
\item{\code{enddate}}{Artist death date}
\item{\code{gender}}{Artist gender}
\item{\code{nationality}}{Artist nationality}
}
}
\description{
Queries the Getty ULAN to find the best matching ID for a given string,
returning the ID as well as various artist attributes. You may filter the
results by specifying an early or late date.
}
\note{
\code{method = "sparql"} requires an internet connection.
}
\examples{
\dontrun{ulan_data("Rembrandt", early_year = 1600,
                 late_year = 1700, method = "sparql")}
\dontrun{ulan_data(c("Rembrandt", "Rothko"), early_year = c(1600, 1900),
                 late_year = c(1700, 2000), method = "sparql")}
}
\seealso{
\link{ulan_id}
}

