% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulanr.R
\name{ulan_id}
\alias{ulan_id}
\title{Return a vector of matching ULAN IDs}
\usage{
ulan_id(names, early_year = -9999, late_year = 2090, inclusive = TRUE,
  method = c("sparql", "local"), max_results = 1)
}
\arguments{
\item{names}{A character vector of names to match to a canonical ULAN id.}

\item{early_year}{Match only artists who died after this year. Like
\code{late_year}, this argument should be a numeric vector of length 1, or
of the same length as \code{names}. If length 1, the same date restrictions
will be used to match every value of \code{names}. Otherwise, each name
match can be restricted to its own pair of early_year and late_year. If no
\code{early_year} or \code{late_year} are specified, then artists from all
time periods will be eligible for matching. Any NA values in
\code{early_year} or \code{late_year} will be coerced to default maxima and
minima.}

\item{late_year}{Match only artists who were born before this year.}

\item{inclusive}{Method for filtering search results using the
early_year/late_year parameters. TRUE (the default) will only include
artists whose life dates fall within the range [late_year, early_year].
FALSE will include artists whose life dates intersect with [early_year,
late_year]}

\item{method}{This value determines which method will be used to match the
name to a canonical ULAN id. \code{sparql} will query the Getty's live
endpoint, relying on its Lucene index for finding close matches, while
\code{local} instead uses string distance measures based on a local table
of ULAN entries.}

\item{max_results}{The maximum number of results to return. Defaults to 5.
Depending on the query, the actual number of results returned may be lower.
When \code{method = "sparql"} is used, values over 50 will be ignored.}
}
\value{
An integer vector of ULAN IDs.
}
\description{
This function is a wrapper around \link{ulan_match} that simply returns a
vector of the top match ID for each name given (as if \code{max_results} were
1).
}
\note{
\code{max_results} will be ignored.
}

