% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulanr.R
\name{ulan_id}
\alias{ulan_id}
\title{Name to ULAN ID}
\usage{
ulan_id(names, early_year = -9999, late_year = 2090, method = c("sparql"))
}
\arguments{
\item{names}{A character vector of names to match to a canonical ULAN id.}

\item{early_year}{Match only artists who died after this year. Like
\code{late_year}, this argument should be a numeric vector of length 1, or
of the same length as \code{names}. If length 1, the same date restrictions
will be used to match every value of \code{names}. Otherwise, each name
match can be restricted to its own pair of early_year and late_year. If no
\code{early_year} or \code{late_year} are specified, then artists from all
time periods will be eligible for matching.}

\item{late_year}{Match only artists who were born before this year.}

\item{method}{This value determines which method will be used to match the
  name to a canonical ULAN id.}
}
\value{
A canonical ULAN id for an artist as a 9-digit integer. If no match
  is found, then NULL will be returned isntead.
}
\description{
Queries the Getty ULAN to find the best matching ID for a given string. You
may filter the results by specifying an early or late date.
}
\note{
\code{method = "sparql"} requires an internet connection.
}
\examples{
\dontrun{ulan_id("Rembrandt", early_year = 1600, late_year = 1700, method = "sparql")}
\dontrun{ulan_id(c("Rembrandt", "Rothko"), early_year = c(1600, 1900), late_year = c(1700, 2000), method = "sparql")}
}

