% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulan_sparql.R
\name{ulan_sparql}
\alias{ulan_sparql}
\title{Iterate the SPARQL method over the provided vectors}
\usage{
ulan_sparql(names, early_year, late_year, progress_bar)
}
\arguments{
\item{early_year}{Match only artists who died after this year.}

\item{late_year}{Match only artists who were born before this year.}

\item{progress_bar}{Display a progress bar for long vectors.}

\item{name}{A character string of an artist's name}
}
\description{
This internal function maps the inputs from the generic \link{ulan_id}
function to the SPARQL implementation
}

